exercise_4/requirements.txt
Flask
pandas
requests
google-cloud-storage